+++ data-raw/generate_palette.R
@@ -14,11 +14,23 @@
 
 # palette.hpp template
 
+# pal <- colourvalues::convert_colours(
+#   grDevices::heat.colors(n = 256)
+# )
+
+n <- 256
+cols <- 4
+pal_name <- "rainbow_hcl"
+
 pal <- colourvalues::convert_colours(
-  grDevices::topo.colors(n = 256)
+  #RColorBrewer::brewer.pal(n, pal_name)
+  #grDevices::rainbow( n = n )
+  #colorRamps::blue2red(n)
+  colorspace::rainbow_hcl(n = n)
 )
+
+pal_name <- tolower(pal_name)
 pal <- pal / 255
-pal_name <- "topo"
 
 ### ------------
 ### palette.hpp
@@ -26,9 +38,10 @@ pal_name <- "topo"
 
 save_dir <- "./inst/include/colourvalues/palettes/"
 
-format_vector <- function(x) {
-  m <- matrix(x, ncol = 4, byrow = TRUE)
-  res <- vector( mode = "character", length = 256 / 4)
+to_matrix <- function(x, cols) {
+  m <- matrix(x, ncol = cols, byrow = TRUE)
+  #m <- x
+  res <- vector( mode = "character", length = n / cols )
   for( i in 1:nrow(m) ) {
     if ( i != nrow(m) ) {
       res[i] <- paste0("    ", paste0(m[i,], collapse = ","), "," , sep = "\n")
@@ -39,9 +52,10 @@ format_vector <- function(x) {
   return( res )
 }
 
-red <- format_vector( pal[, 1] )
-green <- format_vector( pal[, 2] )
-blue <- format_vector( pal[, 3] )
+red <- to_matrix( pal[, 1], cols)
+green <- to_matrix( pal[, 2], cols )
+blue <- to_matrix( pal[, 3], cols )
+
 
 txt <- paste0(
   "#ifndef R_COLOURVALUES_", toupper(pal_name), "_H\n",
@@ -84,10 +98,26 @@ close(fileConn)
 ### ------
 ### palettes.hpp
 ### ------
+fileConn <- file("./inst/include/colourvalues/palettes.hpp")
+txt <- readLines(con = fileConn)
+close(fileConn)
 
-clipr::write_clip(
-  paste0("#include \"colourvalues/palettes/",pal_name,".hpp\"")
-)
+# clipr::write_clip(
+#   paste0("#include \"colourvalues/palettes/",pal_name,".hpp\"")
+# )
+
+## insert new line before #endif
+endline <- which(grepl("#endif", txt))
+
+newVec <- vector("character", length = length(txt) + 1)
+
+newVec[1:(endline-1)] <- txt[1:(endline-1)]
+newVec[endline-1] <- paste0("#include \"colourvalues/palettes/",pal_name,".hpp\"\n")
+newVec[length(newVec)] <- txt[endline]
+
+#cat( paste0(newVec, collapse = "\n") )
+
+cat(paste0(newVec, collapse = "\n"), file = "./inst/include/colourvalues/palettes.hpp", append = FALSE)
 
 
 
@@ -99,7 +129,7 @@ txt <- paste0(
   "} else if ( palette == \"", pal_name, "\" ) {\n",
   "  red = colourvalues::palette::",pal_name,"_red;\n",
   "  green = colourvalues::palette::",pal_name,"_green;\n",
-  "  blue = colourvalues::palette::",pal_name,"_blue;\n"
+  "  blue = colourvalues::palette::",pal_name,"_blue;"
 )
 
 clipr::write_clip(txt)
@@ -115,21 +145,26 @@ capitalise <- function(x) {
 }
 
 txt <- paste0(
+  "\n",
   "#' ", capitalise(pal_name), "\n",
   "#'\n",
   "#' Data Frame of the ", pal_name, " palette\n",
   "#'\n",
   "#' @export\n",
-  pal_name, " <- function() rcpp_", pal_name, "()"
+  pal_name, " <- function() rcpp_", pal_name, "()\n"
 )
 
-clipr::write_clip(txt)
+#clipr::write_clip(txt)
+
+cat(txt, file = "./R/palettes.R", append = TRUE)
+
 
 #### ---------
 ### palettes.cpp
 ### -----------
 
 txt <- paste0(
+  "\n",
   "// [[Rcpp::export]]\n",
   "  Rcpp::DataFrame rcpp_",pal_name,"() {\n",
   "    return Rcpp::DataFrame::create(\n",
@@ -137,8 +172,10 @@ txt <- paste0(
   "   _[\"green\"] = colourvalues::palette::", pal_name, "_green,\n",
   "   _[\"blue\"] = colourvalues::palette::", pal_name, "_blue\n",
   ");\n",
-"}"
+"}\n"
 )
 
-clipr::write_clip(txt)
+# clipr::write_clip(txt)
+
+cat(txt, file = "./src/palettes.cpp", append = TRUE)