+++ NEWS.md
@@ -1,5 +1,7 @@
 # v0.2.1
 
+* `show_colours()` to plot all colour palettes
+* `colour_palettes()` to list all colour palettes
 * fix NA values in summary [issue 39](https://github.com/SymbolixAU/colourvalues/issues/39)
 * alpha now accepts fewer than 5 values when using a vector